servers/azure-openai-deployments-mcp/requirements.txt
mcp>=1.3.0
azure-identity>=1.17.0
azure-mgmt-cognitiveservices>=13.5.0
azure-mgmt-subscription>=3.1.1
